Limit to version 78
From John on atn:
If you intend to keep updating your 3.* branch, you need to limit it to 78.*, so 91 users never get it. The numbers in the version have no meaning, only the release date. So when a TB91 user is on 4.1 and you release a new 3.* branch version - lets say 3.3.0 - and that version is not limited to 78*, your TB91 users will get that!